requirements.txt
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
decorator==5.2.1
dlib==20.0.0
idna==3.10
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
tk==0.1.0
typing_extensions==4.14.0
urllib3==2.5.0
wheel==0.45.1
